+++ src/Flysystem/InnerAdapter.php
@@ -4,16 +4,23 @@ declare(strict_types=1);
 
 namespace Webf\FlysystemFailoverBundle\Flysystem;
 
+use League\Flysystem\Config;
+use League\Flysystem\FileAttributes;
 use League\Flysystem\FilesystemAdapter;
+use Webf\Flysystem\Composite\CompositeFilesystemAdapter;
 
 /**
  * @psalm-type _Options=array{
  *     time_shift?: int
  * }
+ *
+ * @template T of FilesystemAdapter
+ * @template-implements CompositeFilesystemAdapter<T>
  */
-class InnerAdapter
+class InnerAdapter implements CompositeFilesystemAdapter
 {
     /**
+     * @param T        $adapter
      * @param _Options $options
      */
     public function __construct(
@@ -22,13 +29,93 @@ class InnerAdapter
     ) {
     }
 
-    public function getAdapter(): FilesystemAdapter
-    {
-        return $this->adapter;
-    }
-
     public function getTimeShift(): int
     {
         return $this->options['time_shift'] ?? 0;
+    }
+
+    public function fileExists(string $path): bool
+    {
+        return $this->adapter->fileExists($path);
+    }
+
+    public function write(string $path, string $contents, Config $config): void
+    {
+        $this->adapter->write($path, $contents, $config);
+    }
+
+    public function writeStream(string $path, $contents, Config $config): void
+    {
+        $this->adapter->writeStream($path, $contents, $config);
+    }
+
+    public function read(string $path): string
+    {
+        return $this->adapter->read($path);
+    }
+
+    public function readStream(string $path)
+    {
+        return $this->adapter->readStream($path);
+    }
+
+    public function delete(string $path): void
+    {
+        $this->adapter->delete($path);
+    }
+
+    public function deleteDirectory(string $path): void
+    {
+        $this->adapter->deleteDirectory($path);
+    }
+
+    public function createDirectory(string $path, Config $config): void
+    {
+        $this->adapter->createDirectory($path, $config);
+    }
+
+    public function setVisibility(string $path, string $visibility): void
+    {
+        $this->adapter->setVisibility($path, $visibility);
+    }
+
+    public function visibility(string $path): FileAttributes
+    {
+        return $this->adapter->visibility($path);
+    }
+
+    public function mimeType(string $path): FileAttributes
+    {
+        return $this->adapter->mimeType($path);
+    }
+
+    public function lastModified(string $path): FileAttributes
+    {
+        return $this->adapter->lastModified($path);
+    }
+
+    public function fileSize(string $path): FileAttributes
+    {
+        return $this->adapter->fileSize($path);
+    }
+
+    public function listContents(string $path, bool $deep): iterable
+    {
+        return $this->adapter->listContents($path, $deep);
+    }
+
+    public function move(string $source, string $destination, Config $config): void
+    {
+        $this->adapter->move($source, $destination, $config);
+    }
+
+    public function copy(string $source, string $destination, Config $config): void
+    {
+        $this->adapter->copy($source, $destination, $config);
+    }
+
+    public function getInnerAdapters(): iterable
+    {
+        return [$this->adapter];
     }
 }