+++ src/MessageHandler/DeleteDirectoryHandler.php
@@ -22,7 +22,6 @@ class DeleteDirectoryHandler implements MessageHandlerInterface
         $adapter = $this->adaptersLocator
             ->get($message->getFailoverAdapter())
             ->getInnerAdapter($message->getInnerDestinationAdapter())
-            ->getAdapter()
         ;
 
         try {